+++ tools/src/build_repetition_index.cpp
@@ -2,8 +2,8 @@
  * @file build_repetition_index.cpp
  * @author Alberto Casagrande (alberto.casagrande@uniud.it)
  * @brief Builds the repetition index
- * @version 1.2
- * @date 2026-02-06
+ * @version 1.3
+ * @date 2026-07-14
  *
  * @copyright Copyright (c) 2023-2026
  *
@@ -50,7 +50,7 @@ class RepetitionIndexBuilder: public BasicExecutable
     std::string driver_mutations_filename;
     size_t max_rep_per_type;
     size_t max_unit_size;
-    unsigned int rep_type_size;
+    uint8_t rep_type_size;
     bool quiet;
 
     template<typename GENOME_WIDE_POSITION>
@@ -116,7 +116,7 @@ public:
             ("max-motif-size,s", po::value<size_t>(&max_unit_size)->default_value(50),
              "maximum size of the repeated motif")
             ("force-overwrite,f", "force overwriting output file")
-            ("rep-type-bytes,b", po::value<unsigned int>(&rep_type_size)->default_value(4),
+            ("rep-type-bytes,b", po::value<uint8_t>(&rep_type_size)->default_value(4),
              "size of the repetition type (one among 1, 2, 4, or 8)")
 #if WITH_INDICATORS
             ("quiet,q", "disable output messages")
